requirements.txt
# The Refactoring Swarm - Dependencies
# Core LLM and Agent Frameworks
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0

# Code Analysis Tools
pylint>=3.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# Optional: Alternative Agent Frameworks
# crewai>=0.1.0
# autogen>=0.2.0
